Namespace Adjustment
