+++ MetaMorpheus/GUI/MetaDraw/BioPolymerCoverageTabView.xaml.cs
@@ -1,6 +1,5 @@
 ﻿using System.Linq;
 using EngineLayer;
-using GuiFunctions;
 using System.Windows;
 using System.Windows.Controls;
 using GuiFunctions.MetaDraw;
