+++ MetaMorpheus/GUI/MetaDraw/MetaDraw.xaml.cs
@@ -22,6 +22,7 @@ using System.Windows.Controls;
 using System.Windows.Data;
 using System.Windows.Input;
 using System.Windows.Media;
+using GuiFunctions.MetaDraw.BioPolymerCoverage;
 
 namespace MetaMorpheusGUI
 {
