+++ MetaMorpheus/GUI/MetaDraw/Views/BioPolymerCoverageMapView.xaml.cs
@@ -1,6 +1,7 @@
 ﻿using System.Windows;
 using System.Windows.Controls;
 using GuiFunctions;
+using GuiFunctions.MetaDraw;
 
 namespace MetaMorpheusGUI;
