+++ MetaMorpheus/GuiFunctions/MetaDraw/BioPolymerCoverage/BioPolymerCoverageMapViewModel.cs
@@ -1,10 +1,10 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Windows.Media;
 using System.Windows;
+using System.Windows.Media;
 
-namespace GuiFunctions;
+namespace GuiFunctions.MetaDraw;
 
 public class BioPolymerCoverageMapViewModel : BaseViewModel
 {
