+++ MetaMorpheus/GuiFunctions/MetaDraw/BioPolymerCoverage/BioPolymerCoverageType.cs
@@ -1,0 +1,11 @@
+﻿namespace GuiFunctions.MetaDraw;
+
+public enum BioPolymerCoverageType
+{
+    Unique,
+    UniqueMissedCleavage,
+    TandemRepeat,               // Shared within the same biopolymer
+    TandemRepeatMissedCleavage,
+    Shared,                     // Shared between different biopolymers
+    SharedMissedCleavage,
+}
