+++ MetaMorpheus/Test/MetaDraw/BioPolymerTabViewModelTests.cs
@@ -1,5 +1,7 @@
 using System.Collections.Generic;
 using GuiFunctions;
+using GuiFunctions.MetaDraw;
+using GuiFunctions.MetaDraw.BioPolymerCoverage;
 using NUnit.Framework;
 using Omics;
 using Proteomics;
